+++ scripts/deploy/deployCrossChainVerifierWithRequests.ts
@@ -500,7 +500,7 @@ async function main() {
 
   const pathOutputJson = path.join(
     __dirname,
-    `./deploy_cross_chain_verification_with_requests_output_${chainId}_${networkName}.json`,
+    `../deployments_output/deploy_cross_chain_verification_with_requests_output_${chainId}_${networkName}.json`,
   );
   fs.writeFileSync(pathOutputJson, JSON.stringify(outputJson, null, 1));
 }
