+++ scripts/deploy/deployUniversalVerifier.ts
@@ -2,7 +2,7 @@ import fs from "fs";
 import path from "path";
 import { DeployHelper } from "../helpers/DeployHelper";
 import hre, { ethers, network } from "hardhat";
-import { getConfig } from "../helpers/helperUtils";
+import { getConfig, waitNotToInterfereWithHardhatIgnition } from "../helpers/helperUtils";
 
 async function main() {
   const config = getConfig();
@@ -10,39 +10,37 @@ async function main() {
   if (!ethers.isAddress(stateAddress)) {
     throw new Error("STATE_CONTRACT_ADDRESS is not set");
   }
-
-  const validators: ("mtpV2" | "sigV2" | "v3")[] = ["mtpV2", "sigV2", "v3"];
   const deployStrategy: "basic" | "create2" =
     config.deployStrategy == "create2" ? "create2" : "basic";
-  const [signer] = await hre.ethers.getSigners();
+  const [signer] = await ethers.getSigners();
 
   const deployHelper = await DeployHelper.initialize(null, true);
 
-  const validatorsInfo: any = [];
-  for (const v of validators) {
-    const { validator, groth16VerifierWrapper } = await deployHelper.deployValidatorContracts(
-      v,
-      stateAddress,
-      deployStrategy,
-    );
-    validatorsInfo.push({
-      validatorType: v,
-      validator: await validator.getAddress(),
-      groth16verifier: await groth16VerifierWrapper.getAddress(),
-    });
-  }
+  const verifierLib = await deployHelper.deployVerifierLib();
+  const tx = await verifierLib.deploymentTransaction();
+  await waitNotToInterfereWithHardhatIgnition(tx);
+
+  const universalVerifier = await deployHelper.deployUniversalVerifier(
+    undefined,
+    stateAddress,
+    await verifierLib.getAddress(),
+    deployStrategy,
+  );
 
   const chainId = parseInt(await network.provider.send("eth_chainId"), 16);
   const networkName = hre.network.name;
   const pathOutputJson = path.join(
     __dirname,
-    `./deploy_validators_output_${chainId}_${networkName}.json`,
+    `../deployments_output/deploy_universal_verifier_output_${chainId}_${networkName}.json`,
   );
   const outputJson = {
     proxyAdminOwnerAddress: await signer.getAddress(),
-    validatorsInfo,
+    universalVerifier: await universalVerifier.getAddress(),
+    verifierLib: await verifierLib.getAddress(),
+    state: stateAddress,
     network: networkName,
     chainId,
+    deployStrategy,
   };
   fs.writeFileSync(pathOutputJson, JSON.stringify(outputJson, null, 1));
 }
