+++ scripts/deployments_output/deploy_state_output_2442_zkevm-cardona.json
