+++ scripts/maintenance/prepare-contracts-package.js
