+++ scripts/upgrade/identitytreestore/identitytreestore-upgrade.ts
@@ -77,7 +77,7 @@ async function main() {
 
   const pathOutputJson = path.join(
     __dirname,
-    `../../deploy_identity_tree_store_output_${chainId}_${network}.json`,
+    `../../deployments_output/deploy_identity_tree_store_output_${chainId}_${network}.json`,
   );
   const outputJson = {
     proxyAdminOwnerAddress: await proxyAdminOwnerSigner.getAddress(),
