+++ scripts/upgrade/state/state-upgrade.ts
@@ -132,7 +132,7 @@ async function main() {
 
   const pathOutputJson = path.join(
     __dirname,
-    `../../deploy_state_output_${chainId}_${network}.json`,
+    `../../deployments_output/deploy_state_output_${chainId}_${network}.json`,
   );
   const outputJson = {
     proxyAdminOwnerAddress: await proxyAdminOwnerSigner.getAddress(),
