+++ scripts/upgrade/validators/validator-upgrade.ts
@@ -89,7 +89,7 @@ async function main() {
 
   const pathOutputJson = path.join(
     __dirname,
-    `../../deploy_validators_output_${chainId}_${network}.json`,
+    `../../deployments_output/deploy_validators_output_${chainId}_${network}.json`,
   );
   const outputJson = {
     proxyAdminOwnerAddress: await proxyAdminOwnerSigner.getAddress(),
